Move common functions used in test to helper file
