+++ test/convert.ts
@@ -1,188 +1,20 @@
-import test, { ExecutionContext } from 'ava';
-import fs from 'fs';
+import test from 'ava';
 import cloneDeep from 'lodash.clonedeep';
 import Metalsmith from 'metalsmith';
-import path from 'path';
 import pug from 'pug';
 import sinon from 'sinon';
-import slug from 'slug';
-import util from 'util';
 
 import { compile, convert, render } from '../src';
-import { isObject } from '../src/utils';
-
-function objIgnoreKeys<T>(obj: T, keyList: string[]): T {
-    if (isObject(obj)) {
-        // @ts-ignore: TS2322
-        const newObj: T = {};
-        return Object.entries(obj)
-            .filter(([key]) => !keyList.includes(key))
-            .reduce((obj, [key, value]) => {
-                // @ts-ignore: TS7053
-                obj[key] = value;
-                return obj;
-            }, newObj);
-    }
-    return obj;
-}
-
-const destDirSet = new Set();
-function getDestDir(t: ExecutionContext): string {
-    const dirName = slug(t.title, slug.defaults.modes['rfc3986']);
-    if (destDirSet.has(dirName)) {
-        throw new Error(`Duplicate test dir: ${dirName}`);
-    }
-    destDirSet.add(dirName);
-    return dirName;
-}
-
-function createMetalsmith(t: ExecutionContext): Metalsmith {
-    return Metalsmith(path.join(__dirname, 'fixtures'))
-        .source('pages')
-        .destination(path.join('build', getDestDir(t)))
-        .clean(true);
-}
-
-function destPath(metalsmith: Metalsmith, ...paths: string[]): string {
-    return path.join(
-        path.relative(process.cwd(), metalsmith.destination()),
-        ...paths,
-    );
-}
-
-async function readSourceFile(
-    metalsmith: Metalsmith,
-    filename: string,
-): Promise<string> {
-    const readFile = util.promisify(fs.readFile);
-    const sourceFilepath = metalsmith.path(metalsmith.source(), filename);
-    return (await readFile(sourceFilepath)).toString();
-}
-
-function setLocalsPlugin(locals: {
-    [index: string]: unknown;
-}): Metalsmith.Plugin {
-    return (files, metalsmith, done) => {
-        Object.values(files).forEach(file => {
-            Object.assign(file, locals);
-        });
-        done(null, files, metalsmith);
-    };
-}
-
-function assertFileExists(
-    t: ExecutionContext,
-    filepath: string,
-): Promise<void> {
-    return new Promise(resolve => {
-        fs.stat(filepath, err => {
-            t.falsy(err, 'File exist');
-            resolve();
-        });
-    });
-}
-
-function assertFileNotExists(
-    t: ExecutionContext,
-    filepath: string,
-): Promise<void> {
-    return new Promise(resolve => {
-        fs.stat(filepath, err => {
-            t.truthy(err, 'File does not exist');
-            resolve();
-        });
-    });
-}
-
-function assertFileContentsEquals(
-    t: ExecutionContext,
-    filepath: string,
-    contents: string,
-): Promise<void> {
-    return new Promise(resolve => {
-        fs.readFile(filepath, (err, data) => {
-            t.falsy(err, 'No readFile error');
-
-            if (!err) {
-                t.is(data.toString(), contents, filepath);
-            }
-
-            resolve();
-        });
-    });
-}
-
-function assertMetalsmithBuild({
-    t,
-    metalsmith,
-}: {
-    t: ExecutionContext;
-    metalsmith: Metalsmith;
-}): Promise<void> {
-    return new Promise(resolve => {
-        metalsmith.build(err => {
-            t.is(err, null, 'No build error');
-            resolve();
-        });
-    });
-}
-
-async function assertFileConverted({
-    t,
-    metalsmith,
-    sourceFilename,
-    destFilename,
-    destFileContents,
-}: {
-    t: ExecutionContext;
-    metalsmith: Metalsmith;
-    sourceFilename?: string;
-    destFilename: string;
-    destFileContents: string;
-}): Promise<void> {
-    await assertMetalsmithBuild({
-        t,
-        metalsmith,
-    });
-    await assertFileContentsEquals(
-        t,
-        destPath(metalsmith, destFilename),
-        destFileContents,
-    );
-    if (sourceFilename && sourceFilename !== destFilename) {
-        await assertFileNotExists(t, destPath(metalsmith, sourceFilename));
-    }
-}
-
-async function assertFileNotConverted({
-    t,
-    metalsmith,
-    sourceFilename,
-    destFilename,
-}: {
-    t: ExecutionContext;
-    metalsmith: Metalsmith;
-    sourceFilename: string;
-    destFilename: string;
-}): Promise<void> {
-    await assertMetalsmithBuild({
-        t,
-        metalsmith,
-    });
-    await assertFileNotExists(t, destPath(metalsmith, destFilename));
-    await assertFileExists(t, destPath(metalsmith, sourceFilename));
-}
-
-function assertFilesPlugin(
-    t: ExecutionContext,
-    fileNameList: string[],
-    message?: string,
-): Metalsmith.Plugin {
-    return (files, metalsmith, done) => {
-        t.deepEqual(Object.keys(files).sort(), fileNameList.sort(), message);
-        done(null, files, metalsmith);
-    };
-}
+import {
+    assertFileConverted,
+    assertFileNotConverted,
+    assertFilesPlugin,
+    assertMetalsmithBuild,
+    createMetalsmith,
+    objIgnoreKeys,
+    readSourceFile,
+    setLocalsPlugin,
+} from './helpers';
 
 test('should render html: convert()', async t => {
     const metalsmith = createMetalsmith(t).use(convert());
